--- src/app/events/page.tsx
@@ -1,7 +1,0 @@
-function Page() {
-  return (
-    <div>
-      
-    </div>
-  );
-}
